+++ R/LearnerClustSOM.R
@@ -69,15 +69,16 @@ LearnerClustSOM = R6Class(
       pv = remove_named(pv, names(grid_args))
       pv$grid = invoke(kohonen::somgrid, .args = grid_args)
 
-      m = invoke(kohonen::som, X = as.matrix(task$data()), .args = pv)
+      data = as.matrix(task$data())
+      m = invoke(kohonen::som, X = data, .args = pv)
       if (self$save_assignments) {
-        self$assignments = as.integer(m$unit.classif)
+        self$assignments = as.integer(m$unit.classif %??% invoke(kohonen::map, m, newdata = data)$unit.classif)
       }
       m
     },
 
     .predict = function(task) {
-      p = invoke(predict, self$model, newdata = as.matrix(ordered_features(task, self)))
+      p = invoke(kohonen::map, self$model, newdata = as.matrix(ordered_features(task, self)))
       PredictionClust$new(task = task, partition = as.integer(p$unit.classif))
     }
   )