+++ tests/testthat/test_mlr_learners_clust_dbscan_fpc.R
@@ -7,6 +7,13 @@ test_that("autotest", {
   expect_true(result, info = result$error)
 })
 
+test_that("predicting errors informatively when trained with seeds = FALSE", {
+  task = tsk("usarrests")
+  learner = lrn("clust.dbscan_fpc", eps = 25, seeds = FALSE)
+  learner$train(task)
+  expect_snapshot(error = TRUE, learner$predict(task))
+})
+
 test_that("Learner properties are respected", {
   task = tsk("usarrests")
   learner = lrn("clust.dbscan_fpc", eps = 25)